Phase 2: Major ESLint compliance improvements
- Fixed file and directory naming conventions (InstantlyApi → Instantly)
- Removed unused/duplicate node files for cleaner structure
- Fixed parameter collection sorting in AccountParameters and LeadParameters
- Fixed boolean parameter descriptions to start with 'Whether'
- Applied auto-fixes for 21+ ESLint violations
- Updated package.json and tsconfig.json to reflect new structure
- Fixed NodeConnectionType usage after auto-fix reverted changes

Progress: Reduced from 50+ ESLint errors to only 3 remaining
Remaining: 2 large collection sorting issues in CampaignParameters + 1 duplicate cleanup

All functionality preserved while significantly improving code quality.
