+++ index.ts
@@ -1,6 +1,6 @@
 import { INodeTypes, NodeConnectionType } from 'n8n-workflow';
 
-import { InstantlyApi } from './nodes/InstantlyApi/InstantlyApi.node';
+import { InstantlyApi } from './nodes/Instantly/Instantly.node';
 import { InstantlyApi as InstantlyApiCredentials } from './credentials/InstantlyApi.credentials';
 
 // This exports the nodes for n8n to load
