+++ nodes/Instantly/parameters/AccountParameters.ts
@@ -150,39 +150,11 @@ export const accountParameters: INodeProperties[] = [
 		},
 		options: [
 			{
-				displayName: 'First Name',
-				name: 'firstName',
-				type: 'string',
-				default: '',
-				description: 'The first name for the account',
-			},
-			{
-				displayName: 'Last Name',
-				name: 'lastName',
-				type: 'string',
-				default: '',
-				description: 'The last name for the account',
-			},
-			{
 				displayName: 'Daily Limit',
 				name: 'dailyLimit',
 				type: 'number',
 				default: 100,
 				description: 'Daily email sending limit',
-			},
-			{
-				displayName: 'Tracking Domain Name',
-				name: 'trackingDomainName',
-				type: 'string',
-				default: '',
-				description: 'Tracking domain for the account',
-			},
-			{
-				displayName: 'Tracking Domain Status',
-				name: 'trackingDomainStatus',
-				type: 'string',
-				default: '',
-				description: 'Status of the tracking domain',
 			},
 			{
 				displayName: 'Enable Slow Ramp',
@@ -192,11 +164,32 @@ export const accountParameters: INodeProperties[] = [
 				description: 'Whether to enable slow ramp up for sending limits',
 			},
 			{
+				displayName: 'First Name',
+				name: 'firstName',
+				type: 'string',
+				default: '',
+				description: 'The first name for the account',
+			},
+			{
 				displayName: 'Inbox Placement Test Limit',
 				name: 'inboxPlacementTestLimit',
 				type: 'number',
 				default: 10,
 				description: 'The limit for inbox placement tests',
+			},
+			{
+				displayName: 'Last Name',
+				name: 'lastName',
+				type: 'string',
+				default: '',
+				description: 'The last name for the account',
+			},
+			{
+				displayName: 'Remove Tracking Domain',
+				name: 'removeTrackingDomain',
+				type: 'boolean',
+				default: false,
+				description: 'Whether to remove the tracking domain from the account',
 			},
 			{
 				displayName: 'Sending Gap (Minutes)',
@@ -214,14 +207,21 @@ export const accountParameters: INodeProperties[] = [
 				name: 'skipCnameCheck',
 				type: 'boolean',
 				default: false,
-				description: 'Skip CNAME check for tracking domain',
+				description: 'Whether to skip CNAME check for tracking domain',
 			},
 			{
-				displayName: 'Remove Tracking Domain',
-				name: 'removeTrackingDomain',
-				type: 'boolean',
-				default: false,
-				description: 'Whether to remove the tracking domain from the account',
+				displayName: 'Tracking Domain Name',
+				name: 'trackingDomainName',
+				type: 'string',
+				default: '',
+				description: 'Tracking domain for the account',
+			},
+			{
+				displayName: 'Tracking Domain Status',
+				name: 'trackingDomainStatus',
+				type: 'string',
+				default: '',
+				description: 'Status of the tracking domain',
 			},
 			{
 				displayName: 'Warmup Configuration',
@@ -235,8 +235,11 @@ export const accountParameters: INodeProperties[] = [
 						displayName: 'Warmup Limit',
 						name: 'limit',
 						type: 'number',
-						default: 100,
-						description: 'Daily warmup email limit',
+						typeOptions: {
+							minValue: 1,
+						},
+						default: 50,
+						description: 'Max number of results to return',
 					},
 					{
 						displayName: 'Reply Rate',
