+++ package.json
@@ -27,8 +27,8 @@
     "build": "tsc && gulp build:icons",
     "dev": "tsc --watch",
     "format": "prettier nodes credentials --write",
-    "lint": "eslint nodes credentials package.json",
-    "lintfix": "eslint nodes credentials package.json --fix",
+    "lint": "eslint nodes/Instantly nodes/generic.functions.ts credentials",
+    "lintfix": "eslint nodes/Instantly nodes/generic.functions.ts credentials --fix",
     "test": "jest",
     "test:unit": "jest __tests__/unit",
     "test:integration": "jest __tests__/integration",
@@ -53,7 +53,7 @@
       "dist/credentials/InstantlyApi.credentials.js"
     ],
     "nodes": [
-      "dist/nodes/InstantlyApi/InstantlyApi.node.js"
+      "dist/nodes/Instantly/Instantly.node.js"
     ]
   },
   "devDependencies": {
