+++ tsconfig.json
@@ -21,7 +21,7 @@
     "noUnusedParameters": false
   },
   "include": [
-    "nodes/InstantlyApi/**/*.ts",
+    "nodes/Instantly/**/*.ts",
     "nodes/generic.functions.ts",
     "credentials/**/*.ts",
     "index.ts"
